requirements.txt
click==6.6
cssselect==0.9.2
hodorlive==1.1.1
lxml==3.6.4
Pillow==3.4.2
publicsuffix==1.1.0
python-dateutil==2.6.0
reppy==0.3.3
requests==2.11.1
six==1.10.0
url==0.1.7